+++ src/multipole_interactions/multipole_kernel_interface.cpp
@@ -29,6 +29,14 @@
 
 #include "octotiger/options.hpp"
 
+#if defined(OCTOTIGER_HAVE_KOKKOS) && defined(KOKKOS_ENABLE_SYCL)
+#include "octotiger/sycl_initialization_guard.hpp"
+/// Dummy variable to ensure the touch_sycl_device_by_running_a_dummy_kernel is being run
+const int init_sycl_device_multipoles =
+    octotiger::sycl_util::touch_sycl_device_by_running_a_dummy_kernel<
+        "gravity_solver_multipoles">();
+#endif
+
 #ifdef OCTOTIGER_HAVE_KOKKOS
 #if defined(KOKKOS_ENABLE_CUDA)
 using device_executor = hpx::kokkos::cuda_executor;
